refactor(grains): restructure GrainFactoryExtensions into specific Hero-related extensions for better organization
